marketplace_backend/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
eth-keys
eth-utils
eth-hash[pycryptodome]
python-dotenv
